requirements.txt
streamlit>=1.25.0
numpy>=1.24
requests>=2.28
python-dotenv>=1.0
language-tool-python>=2.8.0
pytest>=7.0
google-genai>=0.1.0
pyngrok>=6.0.0
language-tool-python 
vaderSentiment 
lexicalrichness